fix bugs + reorganize code
